+++ lib/ts_api_handler.dart
@@ -1,5 +1,8 @@
 library;
 
+export 'package:api_cache_manager/models/cache_db_model.dart' show APICacheDBModel;
+export 'package:dio/dio.dart' show CancelToken, Response;
+
 export 'src/api_cache_helper.dart' show ApiCacheHelper;
 export 'src/api_service.dart' show ApiServices;
 export 'src/utils/error_source_extension.dart' show ErrorSource;